Move WorkshopForm to dashboards/admin
- Updated import paths for WorkshopForm
