+++ src/app/admin/workshops/[id]/edit/page.tsx
@@ -2,7 +2,7 @@
 
 import { useState, useEffect, useCallback } from "react";
 import { useRouter, useParams } from "next/navigation";
-import { WorkshopForm } from "@/components/admin/forms/WorkshopForm";
+import { WorkshopForm } from "@/components/dashboards/admin/WorkshopForm";
 import { AdminPageLayout } from "@/components/dashboards/admin/shared/AdminPageLayout";
 import { useWorkshops } from "@/hooks/admin/useWorkshops";
 import { AdminWorkshopFormData } from "@/types/admin";
